+++ src/components/ApiPicker.js
@@ -20,12 +20,6 @@ class ApiPicker extends Component {
     this.createPicker = this.createPicker.bind(this);
     this.pickerCallback = this.pickerCallback.bind(this);
     this.onAuthApiLoad = this.onAuthApiLoad.bind(this);
-    this.loadSlidesApi = this.loadSlidesApi.bind(this);
-    this.listSlides = this.listSlides.bind(this);
-  }
-
-  componentWillUnmount(){
-    this.loadSlidesApi();
   }
 
   onApiLoad() {
@@ -89,58 +83,6 @@ class ApiPicker extends Component {
     this.setState({
       picked: true,
     });
-
-  }
-
-  loadSlidesApi() {
-    window.gapi.client.load('slides', 'v1').then(this.listSlides);
-  }
-
-  listSlides() {
-    const presentationId = this.state.presentationId;
-
-    let requests = [];
-    /* requests.push({
-      replaceAllText: {
-        containsText: {
-          text: '{{name}}'
-        },
-        replaceText: this.props.name
-      }
-    });
-    requests.push({
-      replaceAllText: {
-        containsText: {
-          text: '{{email}}'
-        },
-        replaceText: this.props.email
-      }
-    });
-    requests.push({
-      replaceAllText: {
-        containsText: {
-          text: '{{phoneNumber}}'
-        },
-        replaceText: this.props.phoneNumber
-      }
-    }); */
-
-    window.gapi.client.slides.presentations.batchUpdate({
-      presentationId: presentationId,
-      requests: requests
-    }).then((response) => {
-
-      console.log(JSON.stringify(response.result).match(/(?<!{){{\s*[\w\.]+\s*}}(?!})/g));
-      console.log("??????");
-    });
-
-  /*   window.gapi.client.slides.presentations.batchUpdate({
-      presentationId: presentationId,
-      requests: requests
-    }).then((response) => {
-      console.log(response);
-      console.log("??????");
-    }); */
   }
 
   render() {
